Renamed/moved directory cmd to internal
All of the packages under the former cmd directory are really internal
packages, not executables.
