+++ internal/commands/diagnose/authcheck.tmpl
