+++ internal/commands/diagnose/networkcheck.tmpl
