+++ internal/commands/diagnose/networkchecks.go
