+++ internal/commands/initconfig/initconfig_test.go
