+++ internal/commands/initconfig/observe-agent.tmpl
