+++ internal/commands/status/status.tmpl
