+++ internal/commands/status/statusretriever_test.go
