+++ internal/connections/hostmonitoring.go
