+++ internal/connections/selfmonitoring.go
